# service-generator/spring-server/src/main/java/io/swagger/api/InfoApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.annotations.*;
import io.swagger.model.ServiceInformation;
import io.swagger.model.SingleEndpointConfiguration;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-09-20T17:09:45.498Z")

@Api(value = "info", description = "the info API")
public interface InfoApi {

    @ApiOperation(value = "Find specyfication of concrete endpoint", notes = "Returns a single function description", response = SingleEndpointConfiguration.class, tags={ "info", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = SingleEndpointConfiguration.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied", response = Void.class),
        @ApiResponse(code = 404, message = "Endpoint not found", response = Void.class) })
    
    @RequestMapping(value = "/info/func/{funcId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<SingleEndpointConfiguration> funcfuncId(@ApiParam(value = "ID of func to return",required=true ) @PathVariable("funcId") Long funcId);


    @ApiOperation(value = "Returns api specyfication", notes = "Returns all informations about how to run calculations on microservice", response = Void.class, tags={ "info", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class),
        @ApiResponse(code = 400, message = "Invalid status value", response = Void.class) })
    
    @RequestMapping(value = "/info",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<ServiceInformation> infoGet();

}