# service-generator/swagger_definitions/swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is a sample server Petstore server.  You can find out more about    
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,
    #swagger](http://swagger.io/irc/).      For this sample, you can use the api
    key `special-key` to test the authorization     filters.
  version: 1.0.0
  title: Pwr
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: petstore.swagger.io
basePath: /
tags:
  - name: info
    description: Returns base information about application api
schemes:
  - http
paths:
  /info:
    get:
      tags:
        - info
      summary: Returns api specyfication
      description: Returns all informations about how to run calculations on microservice
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            items:
              $ref: '#/definitions/ServiceInformation'
        '400':
          description: Invalid status value
  '/info/func/{funcId}':
    get:
      tags:
        - info
      summary: Find specyfication of concrete endpoint
      description: Returns a single function description
      operationId: 'func/{funcId}'
      produces:
        - application/json
      parameters:
        - name: funcId
          in: path
          description: ID of func to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/SingleEndpointConfiguration'
        '400':
          description: Invalid ID supplied
        '404':
          description: Endpoint not found
definitions:
  ServiceInformation:
    type: object
    required:
      - endpoints
    properties:
      endpoints:
        type: array
        items:
          $ref: '#/definitions/SingleEndpointConfiguration'
  SingleEndpointConfiguration:
    type: object
    required:
      - path
      - method
      - fields
    properties:
      path:
        description: relative path to enpoint
        type: string
      method:
        type: string
        description: request type
        enum:
          - GET
          - PUT
          - POST
      fields:
        description: Detailed filed descriptions
        type: array
        items:
          $ref: '#/definitions/FieldDefinition'
  FieldDefinition:
    type: object
    required:
      - name
      - type
      - required
    properties:
      name:
        description: json field name
        type: string
      type:
        description: Type of field
        type: string
        enum:
          - string
          - int
          - float
          - file
      required:
        description: is field required
        type: boolean
      additional_description:
        type: string
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'